Update auf matrix-sdk 0.18, Rust Edition 2024
matrix-sdk stand auf 0.6.2 (2023). Das Update auf 0.18.0 erzwingt
Edition 2024 und rustc >= 1.93; dafuer gibt es jetzt eine
rust-toolchain.toml, damit lokal, CI und Docker identisch bauen.

Notwendige API-Anpassungen:
- Room ist kein Enum mehr (Room::Joined/Invited/Left), sondern ein Typ
  mit room.state() -> RoomState. Der Typ matrix_sdk::room::Joined ist
  ersatzlos entfallen, alle Signaturen nehmen jetzt &Room.
- Login laeuft ueber client.matrix_auth().login_username(...).
- room.send(content, None) -> room.send(content).
- room.event(id) -> room.event(id, None) und liefert TimelineEvent;
  die Deserialisierung geht ueber .raw() und ergibt ein
  AnySyncTimelineEvent.
- Die Reaktion zeigt in ruma 0.16 direkt ueber relates_to auf das
  annotierte Event. Der Umweg ueber die Relation-Enum aus
  room::encrypted entfaellt.
- rusqlite 0.40 implementiert ToSql nicht mehr fuer u64, der
  Reaktionszeitstempel wird als i64 gebunden.

Weiter aktualisiert: rusqlite 0.29 -> 0.40.2 (die uncommittete
Working-Tree-Aenderung auf 0.30 geht hier auf), anyhow, regex, tokio.
matrix-sdk wird bewusst mit default-features = false eingebunden: der
Bot unterstuetzt keine verschluesselten Raeume, damit entfallen
e2e-encryption und die sqlite-Stores aus dem Abhaengigkeitsbaum.

Nebenbei loest das Update die Build-Warnung zu x25519-dalek 1.2.0
("code that will be rejected by a future version of Rust") auf, die
transitiv ueber die alte SDK-Version kam.

+++ src/data/user_reaction.rs
@@ -27,7 +27,8 @@ impl UserReaction {
             println!("{}", msg);
             return Err(msg.to_string());
         }
-        let epoch_secs = duration_since_epoch_opt.unwrap().as_secs();
+        // rusqlite 0.40 no longer implements ToSql for u64, SQLite integers are i64 anyway.
+        let epoch_secs = duration_since_epoch_opt.unwrap().as_secs() as i64;
 
         conn.execute(
             sql,
